+++ lib/src/bones_api_entity_db_object_memory.dart
@@ -706,7 +706,7 @@ class DBMemoryObjectAdapter extends DBAdapter<DBMemoryObjectAdapterContext> {
           var listEntityType = fieldType.listEntityType!;
 
           var fieldListEntityRepository =
-              getEntityRepositoryByType(listEntityType.type);
+              getEntityRepositoryByTypeInfo(listEntityType);
           if (fieldListEntityRepository == null) {
             throw StateError(
                 "Can't determine `EntityRepository` for field `$key` List type: fieldType=$fieldType");
@@ -719,9 +719,10 @@ class DBMemoryObjectAdapter extends DBAdapter<DBMemoryObjectAdapterContext> {
                   ? fieldListEntityRepository.getEntityID(v)
                   : v)
               .toList();
-        } else if (!fieldType.isPrimitiveType &&
-            EntityHandler.isValidEntityType(fieldType.type)) {
-          var fieldEntityRepository = getEntityRepositoryByType(fieldType.type);
+        } else if (!fieldType.isPrimitiveType && fieldType.entityType != null) {
+          var entityType = fieldType.entityType!;
+          var fieldEntityRepository = getEntityRepositoryByType(entityType);
+
           if (fieldEntityRepository == null) {
             throw StateError(
                 "Can't determine `EntityRepository` for field `$key`: fieldType=$fieldType");