+++ lib/src/bones_api_types.dart
@@ -5,6 +5,8 @@ import 'package:collection/collection.dart';
 import 'package:reflection_factory/reflection_factory.dart';
 import 'package:statistics/statistics.dart';
 
+import 'bones_api_entity.dart';
+
 /// Extension for entity [Type]s.
 extension APIEntityTypeExtension on Type {
   /// Returns `true` if `this` [Type] is an [int] or [String].
@@ -32,6 +34,23 @@ extension APIEntityObjectExtension on Object? {
   bool get isEntityIDType {
     var self = this;
     return isEntityIDPrimitiveType || self is BigInt;
+  }
+
+  /// Returns `true` if `this` object is an [EntityReference].
+  bool get isEntityReference {
+    var self = this;
+    return self is EntityReference;
+  }
+
+  /// Returns an entity instance.
+  /// - If it's an [EntityReference] returns [EntityReference.entity].
+  /// - If [isEntityIDType] returns `this` object.
+  /// - Otherwise returns `null`.
+  Object? get resolveEntityInstance {
+    var self = this;
+    return self is EntityReference
+        ? self.entity
+        : (isEntityIDType ? self : null);
   }
 }