+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.3.7
+
+- Added `EntityReference`:
+  An entity field wrapper that allows lazy load of sub-entities.
+- Expose `EntityResolutionRules` to allow lazy or eager selects.
+
 ## 1.3.6
 
 - Added integration with `AsyncEvent`.